+++ CGOpenGL/Input.h
@@ -30,10 +30,10 @@ public:
 
 	bool IsLMBPressed() const;
 	bool IsRMBPressed() const;
-	double GetInputAxisX() const;
-	double GetInputAxisY() const;
-	double GetMouseDeltaX() const;
-	double GetMouseDeltaY() const;
+	float GetInputAxisX() const;
+	float GetInputAxisY() const;
+	float GetMouseDeltaX() const;
+	float GetMouseDeltaY() const;
 
 private:
 	Input();
@@ -49,8 +49,8 @@ private:
 	double lastMousePosY = 0.0;
 	double mousePosX = 0.0;
 	double mousePosY = 0.0;
-	double mouseDeltaXNorm = 0.0;
-	double mouseDeltaYNorm = 0.0;
+	float mouseDeltaXNorm = 0.0;
+	float mouseDeltaYNorm = 0.0;
 
 	// Hooks
 	std::unordered_multimap<std::pair<int, int>, std::function<void()>> mouseButtonHooks;
